lina_morning_report/requirements.txt
google-generativeai==0.8.3
feedparser==6.0.11
python-dotenv==1.0.1
markdown==3.7
requests==2.31.0
